requirements.txt
Django==1.6.5
Pillow==2.0.0
South==0.8.4
argparse==1.2.1
cmsplugin-filer==0.9.5
dj-database-url==0.3.0
dj-static==0.0.5
django-classy-tags==0.5
django-cms==3.0.3
django-filer==0.9.5
django-mptt==0.6.0
django-polymorphic==0.5.5
django-reversion==1.7
django-sekizai==0.7
django-toolbelt==0.0.1
djangocms-admin-style==0.2.2
djangocms-file==0.0.1
djangocms-flash==0.0.2
djangocms-googlemap==0.0.5
djangocms-inherit==0.0.1
djangocms-link==1.3.4
djangocms-picture==0.0.2
djangocms-snippet==1.0.2
djangocms-teaser==0.0.1
djangocms-text-ckeditor==2.1.6
djangocms-video==0.0.1
easy-thumbnails==2.0.1
gunicorn==19.0.0
html5lib==1.0b1
psycopg2==2.5.3
pystache==0.5.4
six==1.3.0
static==1.0.2
wsgiref==0.1.2
